% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{cat_plot}
\alias{cat_plot}
\title{Plot the plot in plotfun and cat as a base64 encoded <img>}
\usage{
cat_plot(plotfun, ...)
}
\description{
Plot the plot in plotfun and cat as a base64 encoded <img>
}

