% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{clusterinfo}
\alias{clusterinfo}
\title{Create a 'clusterinfo' object that can be used for the individual render methods}
\usage{
clusterinfo(m, terms = NULL, documents = NULL, words = terms,
  meta = NULL)
}
\arguments{
\item{m:}{the fitted LDA model object from the topicmodels package}

\item{terms:}{a vector of terms, which should be sorted in their original order and match m@terms}

\item{documents:}{a vector of the same length as terms, indicating the document of each term, matching m@documents}

\item{words:}{if given, use instead of terms for displaying document.}

\item{meta:}{Optional meta data. Vectors or a dataframe of which the elements/rows match m@documents.}
}
\value{
a list with tokens, wordassignments, and other items
}
\description{
Create a 'clusterinfo' object that can be used for the individual render methods
}

