% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{createTopicBrowser}
\alias{createTopicBrowser}
\title{Create a topic browser}
\usage{
createTopicBrowser(info, plotfunction.overview = pf.wordcloud,
  plotfunction.pertopic = c(pf.wordcloud, pf.topicdistribution), ...,
  output = NULL, nmaxdoc = 10, nmaxwords = 100, browse = interactive(),
  topic_ids = 1:nrow(info$topics_per_doc))
}
\arguments{
\item{info}{The output of the \code{\link{clusterinfo}} function}

\item{plotfunction.overview}{A standardized function (see the `pf.` functions). Determines how the topics in the general topic overview are visualized}

\item{plotfunction.pertopic}{A standardized function (see the `pf.` functions). Determines what information is visualized in the pages per topic. Can be a list of multiple functions, in which case multiple plots are made.}

\item{...}{Parameters to be passed to the plotfunctions.}

\item{output}{an optional file name to write the browser html to}

\item{nmaxdoc}{an integer, indicating the maximum number of top articles to be printed for each topic.}
}
\value{
the html generated or (invisible) the filename
}
\description{
This will create a set of linked html pages to browse the given topic model
}

