% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{getTopDocuments}
\alias{getTopDocuments}
\title{Get top documents for a topic}
\usage{
getTopDocuments(clusterinfo, topic_nrs, ndocs)
}
\arguments{
\item{topic_nr:}{the topic nr to render}

\item{n}{the number of top documents}
}
\description{
Get top documents for a topic
}

