% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{listFunctionArguments}
\alias{listFunctionArguments}
\title{Match elipsis arguments to the permitted arguments of a function}
\usage{
listFunctionArguments(func, ...)
}
\arguments{
\item{func}{the function}

\item{...}{any argument}
}
\value{
a list with all and only given arguments (...) that can be used by the function
}
\description{
Match elipsis arguments to the permitted arguments of a function
}

