% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{pf.wordcloud}
\alias{pf.wordcloud}
\title{Standardized plotting function that can be passed to the plotfunction arguments in the \code{\link{createTopicBrowser}} function.}
\usage{
pf.wordcloud(info, topic_nr)
}
\arguments{
\item{info}{The output of the \code{\link{clusterinfo}} function}

\item{topic_nr}{the index number of a topic}
}
\value{
nothing, only plots
}
\description{
One of the standardized plotting function used in the Topicbrowser package to manage how topics and additional information are visualized.
This specific function plots a wordcloud
}

