% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_semnet}
\alias{plot_semnet}
\title{Standardized plotting function that can be passed to the plotfunction arguments in the \code{\link{createTopicBrowser}} function.}
\usage{
plot_semnet(clusterinfo, topic_nr, backbone_alpha = 0.05, nwords = 100)
}
\arguments{
\item{info}{The output of the \code{\link{clusterinfo}} function}

\item{topic_nr}{the index number of a topic}
}
\value{
nothing, only plots
}
\description{
One of the standardized plotting function used in the Topicbrowser package to manage how topics and additional information are visualized.
This specific function plots a wordcloud below a graph showing the number of words assigned to the topic per time period
}

