% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.r
\name{plot_time}
\alias{plot_time}
\title{Standardized plotting function that can be passed to the plotfunction arguments in the \code{\link{createTopicBrowser}} function.}
\usage{
plot_time(lda_model = NULL, document_topic_matrix = NULL,
  clusterinfo = NULL, topic_nr, time_var = "date", time_interval = "auto",
  pct = F, value = "total")
}
\arguments{
\item{topic_nr}{the index number of a topic}

\item{time_var}{either a vector containing the date for each document, or the column name of the date variable in the metadata stored in the clusterinfo object}

\item{time_interval}{a string indicating what interval to use for plotting the time/date graph. Can be 'year', 'month', 'week' and 'day'. If 'auto', the date_interval will be chosen based on the number of days in the analysis.}

\item{info}{The output of the \code{\link{clusterinfo}} function}
}
\value{
nothing, only plots
}
\description{
One of the standardized plotting function used in the Topicbrowser package to manage how topics and additional information are visualized.
This specific function plots a graph showing the number of words assigned to the topic per time period
}

