% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_to_file}
\alias{plot_to_file}
\title{Plot the plot in plotfun to a tempfile and return the filename}
\usage{
plot_to_file(plotfun, width = 500, height = width)
}
\description{
Plot the plot in plotfun to a tempfile and return the filename
}

