% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Function to plot a word cloud for a given topic}
\usage{
plot_wordcloud(lda_model = NULL, topic_term_matrix = NULL,
  clusterinfo = NULL, topic_nr, wordsize_scale = 0.75,
  relative_to_term_total = F)
}
\arguments{
\item{lda_model}{An lda model fitted using topicmodels::LDA}

\item{topic_term_matrix}{a matrix of topics per term}

\item{clusterinfo}{The output of the \code{\link{clusterinfo}} function}

\item{topic_nr}{the index number of a topic}

\item{wordsize_scale}{a scale for the word sizes}

\item{relative_to_term_total}{make word sizes relative}
}
\value{
nothing, only plots
}
\description{
Can be passed to the plotfunction arguments in the \code{\link{createTopicBrowser}} function.
}
\details{
One of the standardized plotting function used in the Topicbrowser package to manage how topics and additional information are visualized.
This specific function plots a wordcloud

One of lda_model, topic_term_matrix, or topicinfo needs to be specified
}

