% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_wordcloud_time}
\alias{plot_wordcloud_time}
\title{Standardized plotting function that can be passed to the plotfunction arguments in the \code{\link{createTopicBrowser}} function.}
\usage{
plot_wordcloud_time(clusterinfo, topic_nr, ...)
}
\arguments{
\item{info}{The output of the \code{\link{clusterinfo}} function}

\item{topic_nr}{the index number of a topic}

\item{date_interval}{if not NULL, a string indicating what interval to use for plotting the time/date graph. Can be 'year', 'month', 'week' and 'day'. If NULL, the date_interval will be chosen based on the number of days in the analysis.}

\item{date_label}{a string indicating what the date object is in the meta data.}
}
\value{
nothing, only plots
}
\description{
One of the standardized plotting function used in the Topicbrowser package to manage how topics and additional information are visualized.
This specific function plots a wordcloud below a graph showing the number of words assigned to the topic per time period
}

