% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_html}
\alias{render_html}
\title{Render the tabbed html containing overview and individual tabs}
\usage{
render_html(clusterinfo, topic_nrs, plotfunction.overview,
  plotfunction.pertopic, nmaxdoc, nmaxwords, ...)
}
\arguments{
\item{info}{the cluster_info object (list)}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

