% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_overview}
\alias{render_overview}
\title{Render the index page}
\usage{
render_overview(topics_per_term, topics_per_doc, meta, topic_ids,
  date_interval = "year")
}
\arguments{
\item{topics_per_term:}{a matrix of terms by topics}

\item{topics_per_doc:}{a matrix of documents by topics}

\item{meta:}{a data frame with meta information about the documents}

\item{topic_ids:}{a vector of topic ids to render}

\item{date_interval:}{optional date interval to use for time graphs}
}
\value{
a raw html string (character)
}
\description{
Render the index page
}

