% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{render_topic}
\alias{render_topic}
\title{Render a single topic}
\usage{
render_topic(clusterinfo, topic_nr, plotfunction, nmaxdoc, nmaxwords, ...)
}
\arguments{
\item{topic_id:}{the topic id to render}

\item{info}{the cluster_info object (list)}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

