% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topics.per.document}
\alias{topics.per.document}
\title{Get a dataframe of topics per docment, with an identifier column}
\usage{
topics.per.document(lda_model = NULL, clusterinfo = NULL)
}
\arguments{
\item{lda_model}{An lda model fitted using topicmodels::LDA}

\item{clusterinfo}{The output of the \code{\link{clusterinfo}} function}
}
\description{
Get a dataframe of topics per docment, with an identifier column
}

