% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topics.plot.category}
\alias{topics.plot.category}
\title{Plots topic values per category}
\usage{
topics.plot.category(document_sums, topic_nr, category_var, pct = T,
  value = "relative")
}
\description{
Plots topic values per category
}

