% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topics_per_document}
\alias{topics_per_document}
\title{Get a dataframe of topics per docment, with an identifier column}
\usage{
topics_per_document(lda_model = NULL, clusterinfo = NULL,
  use.posterior = F)
}
\arguments{
\item{lda_model}{An lda model fitted using topicmodels::LDA}

\item{clusterinfo}{The output of the \code{\link{clusterinfo}} function}

\item{posterior}{use the posterior distribution or the actual word assignment (default)}
}
\description{
Get a dataframe of topics per docment, with an identifier column
}

