% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topicbrowser.r
\name{topicsPerTerm}
\alias{topicsPerTerm}
\title{Extract a topic X term matrix from a topicmodels object}
\usage{
topicsPerTerm(m)
}
\arguments{
\item{m}{a topic model in the format of the topicmodels package.}
}
\value{
a topic X term matrix
}
\description{
Extract a topic X term matrix from a topicmodels object
}

