% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wrap_html}
\alias{wrap_html}
\title{Wrap a call that 'cats' html fragments and save it as a html file with header and footer}
\usage{
wrap_html(wrapped, output = NULL)
}
\arguments{
\item{wrapped}{the wrapped call}

\item{output}{an optional filename to output to. If NULL, a tempfile will be used}
}
\value{
the output filename
}
\description{
Since R evaluates arguments lazily, call with the actual wrapped call, i.e. wrap_html(render_html(...))
}

