% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createTopicBrowser}
\alias{createTopicBrowser}
\title{Create a topic browser}
\usage{
createTopicBrowser(m, terms, documents, meta, folder_name = NULL,
  topic_ids = 1:m@k, date_interval = "year", words = terms)
}
\arguments{
\item{m:}{the fitted LDA model object from the topicmodels package}

\item{terms:}{a vector of terms, which should be sorted in their original order and match m@terms}

\item{documents:}{a vector of the same length as terms, indicating the document of each term, matching m@documents}

\item{meta:}{a data frame with meta data about the documents, should have columns aid, headline, medium, and date (todo: make more flexible)}

\item{folder_name:}{the name of the folder to save the browser to}

\item{topic_ids:}{optionally restrict output to a selection of topics}

\item{date_interval:}{specify the interval for plotting the meta$data}

\item{words:}{if given, use instead of terms for displaying document}
}
\value{
(invisible) the file url of the index page
}
\description{
This will create a set of linked html pages to browse the given topic model
}

