% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_article}
\alias{render_article}
\title{Render a single article}
\usage{
render_article(terms, topics, meta, fragment.only = T)
}
\arguments{
\item{terms:}{a vector of words}

\item{topics:}{a vector containing the topic of each word}

\item{meta:}{a 1-row data frame containing meta information}

\item{fragment.only:}{passed to knit2html (if T, do not output html header etc)}
}
\value{
a raw html character vector
}
\description{
Render a single article
}

