% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_overview}
\alias{render_overview}
\title{Render the index page}
\usage{
render_overview(clusterinfo, topic_nrs, plotfunction, ...)
}
\arguments{
\item{info}{the cluster_info object (list)}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

