% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_topic}
\alias{render_topic}
\title{Render a single topic}
\usage{
render_topic(topic_id, info, nmaxdoc = 10)
}
\arguments{
\item{topic_id:}{the topic id to render}

\item{info}{the cluster_info object (list)}
}
\description{
This function does not return the html, but rather cats it directly
So you might wish to use wrap_html or sink to capture the output
}

