% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sotu}
\alias{sotu}
\title{An LDA topicmodel, tokens and meta for State of the Union speeches}
\format{sotu: A LDA topicmodel from the topicmodels package, a dataframe containing tokens for State of the Union speeches and a dataframe containing the metadata.}
\usage{
data(sotu)
}
\description{
An LDA topicmodel, tokens and meta for State of the Union speeches
}
\keyword{datasets}

