% Generated by roxygen2 (4.0.1): do not edit by hand
\name{topicsPerDoc}
\alias{topicsPerDoc}
\title{Extract a topic X document matrix from a topicmodels object}
\usage{
topicsPerDoc(m)
}
\arguments{
\item{m}{a topic model in the format of the topicmodels package.}
}
\value{
a topic X document matrix
}
\description{
Extract a topic X document matrix from a topicmodels object
}

